Prototyping Viewport class. Created 'abstract' Entity class.
Developing the basics of the Viewport class. For it to work,
it must be passed objects to be rendered which have
a certain 'interface.'

The properties and methods required
have been prototyped in the Entity class, which is intended
to serve as a sort of Abstract Class.
May create a more robust Abstract Class like so:
https://coderwall.com/p/r47jlw/advanced-javascript-class-abstract-class-method

main.js is mainly an empty template for p5.js code.